+++ CHANGELOG.md
@@ -25,7 +25,7 @@ Versions with only mechanical changes will be omitted from the following list.
 * Add support for microseconds timestamps serde serialization for `NaiveDateTime`.
 * Add support for optional timestamps serde serialization for `NaiveDateTime`.
 * Fix build for wasm32-unknown-emscripten (@yu-re-ka #593)
-* Use `libtzfile` to get local offset on unix
+* Use `tz-rs` to get local offset on unix
 
 ## 0.4.19